feat: Kafka 설정 및 초기화 클래스 추가
Kafka 설정과 토픽 초기화 로직을 구현한다.
